pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md-doc-pipeline"
version = "0.1.0"
description = "Markdown → PDF/DOCX document pipeline with cascading config and sync"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["markdown", "pdf", "docx", "document", "pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1.5",
    "weasyprint>=60.0",
    "python-docx>=1.1",
    "markdown>=3.5",
    "click>=8.1",
    "markitdown[pdf,docx]>=0.1.0",
]

[project.optional-dependencies]
azure = ["azure-storage-file-share>=12.0"]
s3 = ["boto3>=1.34"]

[project.scripts]
md-doc = "md_doc.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["md_doc"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    "types-pyyaml>=6.0.12.20260408",
    "types-markdown>=3.10.2.20260408",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = [
    "md_doc/sync/azure_files\\.py",
    "md_doc/sync/s3\\.py",
    "md_doc/builders/docx\\.py",
    "md_doc/builders/dotx\\.py",
]
